backend/requirements.txt
torch
torchvision
pillow
numpy
matplotlib
opencv-python